requirements.txt
# FaceVault — Python dependencies
# Install with:  pip install -r requirements.txt

streamlit>=1.33.0
deepface>=0.0.93
faiss-cpu>=1.8.0        # use faiss-gpu if CUDA is available
numpy>=1.24.0
opencv-python-headless>=4.9.0.80
